docs(PPDSC-2299): fixed background and reorganise files
