+++ site/components/homepage/hero/index.ts
@@ -1,0 +1,1 @@
+export * from './hero';
