+++ site/components/homepage/hero/types.ts
@@ -1,0 +1,3 @@
+export interface HeroProps {
+  themeMode: string;
+}
